Skip building on Windows with Python 3.5
Both Python 3.5 and Python 3.6 use the same VC version on Windows. As
this package doesn't have Python portions, we need only build once for
this VC version. As Python 3.5 will be phased out before Python 3.6 is
phased out, we go ahead and skip building with Python 3.5 on Windows.
This should lighten the burden on AppVeyor a little as well.